Move data fetching to layout & add tests
Move data fetching to layout

Currently data fetching is happening on index page, so the character items
are available there and to the modals that open from there
(e.g. create/delete modals). But data fetching is not happening on
create/delete pages.

This moves data fetching to their parent 'layout' component so that it is
available to every page

Add tests & minor refactoring

Mainly we're adding e2e tests that support the previous feature (i.e.
create/delete page rendering)
Plus minor refactoring of previous tests, with fixes for tests that broke
due to it